09-03-2023
Declare an Empty  .
Take Single Integer  as input.
For next  Lines format 
case .  the  of the  in a separate line.
case .  an element from the queue. If the queue is empty then print  in a separate line.
case .  Integer  to the  .
case 4.  an element at the  of the . If queue is empty print  in a seperate line.
Input Format

Single Integer .
 Lines of format .
Constraints

 <=  <= 
 <=  <= 
Output Format

According to case Print the information explained, if its not possible then -1.
Sample Input 0

5
1
2
3 9
4
1
Sample Output 0

0
-1
9
1
Explanation 0

Q = [ ], size = 0
Q = [ ], can't remove so invalid move -1
Q = [ ]  Q = [9]
Q = [9] so Q.top() = 9
Q = [9] so Q.size() = 1